chore: relocate Support Page and Username Page to common/mypage directory
