+++ src/pages/index.ts
@@ -27,8 +27,8 @@ export { default as AccountSettingsPage } from './user/my/AccountSettingsPage';
 export { default as DeleteAccountPage } from './user/my/DeleteAccountPage';
 export { default as NicknamePage } from './user/my/NicknamePage';
 export { default as NotificationSettingsPage } from './user/my/NotificationSettingsPage';
-export { default as SupportPage } from './user/my/SupportPage';
-export { default as UsernamePage } from './user/my/UsernamePage';
+export { default as SupportPage } from './common/mypage/SupportPage';
+export { default as UsernamePage } from './common/mypage/UsernamePage';
 
 export { default as SellerProfilePage } from './user/market/SellerProfile';
